# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Solar Installation CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: assignedSalesPersonId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: followUpDate
          in: query
          schema:
            type: string
        - name: followUpStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/followup:
    patch:
      operationId: updateLeadFollowup
      tags: [leads]
      summary: Update lead follow-up
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadFollowupUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}/convert:
    post:
      operationId: convertLead
      tags: [leads]
      summary: Convert lead to project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadConvertInput"
      responses:
        "201":
          description: Project created from lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /leads/{id}/notes:
    get:
      operationId: listLeadNotes
      tags: [leads]
      summary: Get notes for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadNote"
    post:
      operationId: createLeadNote
      tags: [leads]
      summary: Add a note to a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadNoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadNote"
  /leads/{id}/timeline:
    get:
      operationId: getLeadTimeline
      tags: [leads]
      summary: Get activity timeline for a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

  # ─── PROJECTS ────────────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: assignedEngineerId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Projects list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FINANCE ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [admin, sales, finance, warehouse, engineer]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, email, role, isActive, createdAt, updatedAt]
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        mobileNumber:
          type: string
        alternateNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        leadSource:
          type: ["string", "null"]
        stage:
          type: string
        stageUpdatedAt:
          type: string
        assignedSalesPersonId:
          type: ["integer", "null"]
        assignedSalesPerson:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        remarks:
          type: ["string", "null"]
        followUpDate:
          type: ["string", "null"]
        followUpStatus:
          type: string
        convertedProjectId:
          type: ["integer", "null"]
        convertedAt:
          type: ["string", "null"]
        projectCode:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerName, mobileNumber, stage, followUpStatus, createdAt, updatedAt]
    LeadInput:
      type: object
      properties:
        customerName:
          type: string
        mobileNumber:
          type: string
        alternateNumber:
          type: string
        address:
          type: string
        city:
          type: string
        email:
          type: string
        leadSource:
          type: string
        stage:
          type: string
        assignedSalesPersonId:
          type: integer
        remarks:
          type: string
        followUpDate:
          type: string
        followUpStatus:
          type: string
      required: [customerName, mobileNumber]
    LeadUpdate:
      type: object
      properties:
        customerName:
          type: string
        mobileNumber:
          type: string
        alternateNumber:
          type: string
        address:
          type: string
        city:
          type: string
        email:
          type: string
        leadSource:
          type: string
        stage:
          type: string
        assignedSalesPersonId:
          type: ["integer", "null"]
        remarks:
          type: string
        followUpDate:
          type: string
        followUpStatus:
          type: string
    LeadFollowupUpdate:
      type: object
      properties:
        followUpDate:
          type: string
        followUpStatus:
          type: string
      required: [followUpDate]
    LeadConvertInput:
      type: object
      properties:
        systemCapacityKw:
          type: number
        totalAmount:
          type: number
        assignedEngineerId:
          type: integer
        address:
          type: string
      required: []
    LeadNote:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        note:
          type: string
        createdById:
          type: ["integer", "null"]
        createdBy:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        createdAt:
          type: string
      required: [id, leadId, note, createdAt]
    LeadNoteInput:
      type: object
      properties:
        note:
          type: string
        createdById:
          type: integer
      required: [note]
...
    Project:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        systemCapacityKw:
          type: ["string", "null"]
        totalAmount:
          type: ["string", "null"]
        stage:
          type: string
        assignedEngineerId:
          type: ["integer", "null"]
        assignedEngineer:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        remarks:
          type: ["string", "null"]
        stageUpdatedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, customerName, customerPhone, stage, createdAt, updatedAt]
    ProjectInput:
      type: object
      properties:
        leadId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        city:
          type: string
        systemCapacityKw:
          type: number
        totalAmount:
          type: number
        stage:
          type: string
        assignedEngineerId:
          type: integer
        remarks:
          type: string
      required: [customerName, customerPhone]
    ProjectUpdate:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        city:
          type: string
        systemCapacityKw:
          type: number
        totalAmount:
          type: number
        stage:
          type: string
        assignedEngineerId:
          type: ["integer", "null"]
        remarks:
          type: string
...
    Activity:
      type: object
      properties:
        id:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        action:
          type: string
        description:
          type: string
        performedById:
          type: ["integer", "null"]
        performedBy:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, entityType, entityId, action, description, createdAt]
...