# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Solar Installation CRM API
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  sales: 'sales',
  finance: 'finance',
  warehouse: 'warehouse',
  engineer: 'engineer',
} as const;